lib/agent-core/docker/requirements.txt
strands-agents==1.21.0
bedrock-agentcore==1.1.4
pydantic==2.12.3
boto3==1.42.26
aws-opentelemetry-distro==0.14.1
